Combined function of GripperBuilder.py and GripperCollector.py into single file gripper_collector.py
